NavigationBar refactored for use with modular CSS and converted to TailwindCSS
