+++ src/components/index.tsx
@@ -1,1 +1,1 @@
-export {default as NavigationBar} from './NavigationBar';
+export { default as NavigationBar } from "./NavigationBar/NavigationBar";
